Move Sql Statement Generation Functionality
