+++ DatabaseLibraryNet/SqlWriter.cs
@@ -391,6 +391,71 @@ namespace DigitalZenWorks.Database.ToolKit
 		}
 
 		/// <summary>
+		/// Generates a SQL CREATE TABLE statement for the specified table
+		/// definition.
+		/// </summary>
+		/// <remarks>The generated SQL statement includes all columns in ordinal
+		/// position order, as well as primary key and foreign key constraints
+		/// if specified in the table definition. The output uses double quotes
+		/// for table and column names to ensure compatibility with the ANSI
+		/// standard syntax.</remarks>
+		/// <param name="table">The table structure containing column
+		/// definitions, primary key, and foreign keys to be used in the
+		/// generated SQL statement. Cannot be null.</param>
+		/// <returns>A string containing the SQL CREATE TABLE statement that
+		/// defines the table, its columns, primary key, and foreign key
+		/// constraints.</returns>
+		public virtual string GetCreateTableStatement(Table table)
+		{
+			ArgumentNullException.ThrowIfNull(table);
+
+			string sql = string.Format(
+				CultureInfo.InvariantCulture,
+				"CREATE TABLE `{0}` ({1}",
+				table.Name,
+				Environment.NewLine);
+
+			SortedList<int, Column> columns = GetOrdinalSortedColumns(table);
+
+			foreach (KeyValuePair<int, Column> entry in columns)
+			{
+				Column column = entry.Value;
+
+				sql += GetColumnSql(column);
+			}
+
+			if (!string.IsNullOrWhiteSpace(table.PrimaryKey))
+			{
+				sql += string.Format(
+					CultureInfo.InvariantCulture,
+					"\tCONSTRAINT PrimaryKey PRIMARY KEY (\"{0}\"),{1}",
+					table.PrimaryKey,
+					Environment.NewLine);
+			}
+
+			for (int index = 0; index < table.ForeignKeys.Count; index++)
+			{
+				ForeignKey foreignKey = table.ForeignKeys[index];
+
+				bool isLast = false;
+
+				if (index == table.ForeignKeys.Count - 1)
+				{
+					isLast = true;
+				}
+
+				sql += GetForeignKeySql(foreignKey, isLast);
+			}
+
+			sql += string.Format(
+				CultureInfo.InvariantCulture,
+				"{0});{0}",
+				Environment.NewLine);
+
+			return sql;
+		}
+
+		/// <summary>
 		/// Get date.
 		/// </summary>
 		/// <param name="column">The column index to check.</param>
@@ -443,6 +508,223 @@ namespace DigitalZenWorks.Database.ToolKit
 				defaultValue);
 
 			return dataItem;
+		}
+
+		/// <summary>
+		/// Returns the SQL type declaration string corresponding to the
+		/// specified column's type and length.
+		/// </summary>
+		/// <remarks>The returned string includes the appropriate SQL type and,
+		/// for string columns, the length constraint. This method does not
+		/// validate the column's properties; callers should ensure that the
+		/// column is properly configured before calling.</remarks>
+		/// <param name="column">The column for which to generate the SQL type
+		/// declaration. The column's type and length determine the returned
+		/// string.</param>
+		/// <returns>A string representing the SQL type declaration for the
+		/// column. Returns an empty string if the column type is not
+		/// recognized.</returns>
+		protected static string GetColumnTypeText(Column column)
+		{
+			ArgumentNullException.ThrowIfNull(column);
+
+			string columnType = column.ColumnType switch
+			{
+				ColumnType.Number => " INTEGER",
+				ColumnType.AutoNumber => " INTEGER",
+				ColumnType.String => string.Format(
+					CultureInfo.InvariantCulture,
+					" VARCHAR({0})",
+					column.Length),
+				ColumnType.Memo => " MEMO",
+				ColumnType.DateTime => " DATETIME",
+				ColumnType.Currency => " CURRENCY",
+				ColumnType.Ole => " OLEOBJECT",
+				ColumnType.YesNo => " OLEOBJECT",
+				_ => string.Empty,
+			};
+
+			return columnType;
+		}
+
+		/// <summary>
+		/// Returns a sorted list of columns from the specified table, ordered
+		/// by their ordinal position.
+		/// </summary>
+		/// <remarks>The returned list is sorted in ascending order by column
+		/// position. If the table contains no columns, the returned list will
+		/// be empty.</remarks>
+		/// <param name="table">The table containing the columns to be sorted.
+		/// Cannot be null.</param>
+		/// <returns>A SortedList where each key is the ordinal position of a
+		/// column and each value is the corresponding Column object from the
+		/// table.</returns>
+		protected static SortedList<int, Column> GetOrdinalSortedColumns(
+			Table table)
+		{
+			ArgumentNullException.ThrowIfNull(table);
+
+			// Sort Columns into ordinal positions
+			SortedList<int, Column> columns = [];
+
+			foreach (KeyValuePair<string, Column> entry in table.Columns)
+			{
+				Column column = entry.Value;
+				columns.Add(column.Position, column);
+			}
+
+			return columns;
+		}
+
+		/// <summary>
+		/// Generates the SQL definition string for the specified column,
+		/// including its name, type, constraints, and default value.
+		/// </summary>
+		/// <remarks>The generated SQL includes the column name, type, and
+		/// applicable constraints such as UNIQUE, NOT NULL, IDENTITY, and
+		/// DEFAULT. The output is intended for use in table creation scripts
+		/// and ends with a comma and newline.</remarks>
+		/// <param name="column">The column for which to generate the SQL
+		/// definition. Cannot be null.</param>
+		/// <returns>A string containing the SQL definition for the column,
+		/// formatted for inclusion in a CREATE TABLE statement.</returns>
+		protected virtual string GetColumnSql(Column column)
+		{
+			ArgumentNullException.ThrowIfNull(column);
+
+			string sql = "\t\"" + column.Name + "\"";
+
+			string columnType = GetColumnTypeText(column);
+			sql += columnType;
+
+			if (column.Unique)
+			{
+				sql += " UNIQUE";
+			}
+
+			if (!column.Nullable)
+			{
+				sql += " NOT NULL";
+			}
+
+			if (column.ColumnType == ColumnType.AutoNumber)
+			{
+				sql += " IDENTITY";
+			}
+
+			if (!string.IsNullOrWhiteSpace(column.DefaultValue))
+			{
+				sql += " DEFAULT " + column.DefaultValue;
+			}
+
+			sql += "," + Environment.NewLine;
+
+			return sql;
+		}
+
+		/// <summary>
+		/// Generates the SQL definition string for the specified column,
+		/// including its name, type, constraints, and default value.
+		/// </summary>
+		/// <remarks>The generated SQL includes the column name, type, and
+		/// applicable constraints such as UNIQUE, NOT NULL, IDENTITY, and
+		/// DEFAULT. The output is intended for use in table creation scripts
+		/// and ends with a comma and newline.</remarks>
+		/// <param name="column">The column for which to generate the SQL
+		/// definition. Cannot be null.</param>
+		/// <returns>A string containing the SQL definition for the column,
+		/// formatted for inclusion in a CREATE TABLE statement.</returns>
+		protected virtual string GetCreateColumnSql(Column column)
+		{
+			ArgumentNullException.ThrowIfNull(column);
+
+			string sql = "\t\"" + column.Name + "\"";
+
+			string columnType = GetColumnTypeText(column);
+			sql += columnType;
+
+			if (column.Unique)
+			{
+				sql += " UNIQUE";
+			}
+
+			if (!column.Nullable)
+			{
+				sql += " NOT NULL";
+			}
+
+			if (column.ColumnType == ColumnType.AutoNumber)
+			{
+				sql += " IDENTITY";
+			}
+
+			if (!string.IsNullOrWhiteSpace(column.DefaultValue))
+			{
+				sql += " DEFAULT " + column.DefaultValue;
+			}
+
+			sql += "," + Environment.NewLine;
+
+			return sql;
+		}
+
+		/// <summary>
+		/// Generates the SQL statement for a foreign key constraint based on
+		/// the specified foreign key definition.
+		/// </summary>
+		/// <remarks>The generated SQL includes ON DELETE CASCADE and ON UPDATE
+		/// CASCADE clauses if the corresponding options are set in the foreign
+		/// key definition. The output is formatted for inclusion in a CREATE
+		/// TABLE statement.</remarks>
+		/// <param name="foreignKey">The foreign key definition containing the
+		/// constraint name, parent column, child table, child column, and
+		/// cascade options. Cannot be null.</param>
+		/// <param name="isLast">Indicates whether this constraint is the last
+		/// in the list. If <see langword="false"/>, a comma is appended to the
+		/// SQL statement.</param>
+		/// <returns>A string containing the SQL statement for the foreign key
+		/// constraint, including cascade options if specified.</returns>
+		protected virtual string GetForeignKeySql(
+			ForeignKey foreignKey, bool isLast)
+		{
+			ArgumentNullException.ThrowIfNull(foreignKey);
+
+			const string constraint = "CONSTRAINT";
+			const string key = "FOREIGN KEY";
+			const string references = "REFERENCES";
+
+			const string statement =
+				"\t{0} \"{1}\" {2} (\"{3}\") {4} \"{5}\" (\"{6}\")";
+
+			string sql = string.Format(
+				CultureInfo.InvariantCulture,
+				statement,
+				constraint,
+				foreignKey.Name,
+				key,
+				foreignKey.ParentColumn,
+				references,
+				foreignKey.ChildTable,
+				foreignKey.ChildColumn);
+
+			if (foreignKey.CascadeOnDelete)
+			{
+				sql += " ON DELETE CASCADE";
+			}
+
+			if (foreignKey.CascadeOnUpdate)
+			{
+				sql += " ON UPDATE CASCADE";
+			}
+
+			if (isLast == false)
+			{
+				sql += ",";
+			}
+
+			sql += Environment.NewLine;
+
+			return sql;
 		}
 	}
 }
